requirements.txt
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.3.0
braintree==4.5.0
cairocffi==1.2.0
CairoSVG==2.5.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.4
django-celery-results==2.0.0
django-debug-toolbar==3.2
flower==0.9.7
html5lib==1.1
humanize==3.2.0
idna==2.10
kombu==4.6.11
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
pycparser==2.20
Pyphen==0.10.0
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
tinycss2==1.1.0
tornado==6.1
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
WeasyPrint==52.2
webencodings==0.5.1
